Update index.html
The dashboard isn't loading, and click events aren't working because of a critical error in the code. My apologies for this. The function responsible for building the "Edit Plan" view (renderEditGrid) was incomplete. This caused a chain reaction: Since the edit grid couldn't be built, the app would get stuck and not display the dashboard data correctly. Because the meal inputs and buttons were never created, none of the click actions (like for ingredients or copy/paste) could be attached. I have now rewritten the affected functions to be complete and correct. This version will properly build all the views and ensure every button is fully functional.